Trigger Cubit and States For Movie Details And Movie Recommendation With Ui

+++ lib/Features/Movies/data/models/movie_details_model.dart
@@ -19,6 +19,6 @@ class MovieDetailsModel extends MovieDetailsEntity {
       releaseDate: json["release_date"],
       runtime: json["runtime"],
       title: json["title"],
-      voteAverage: json["vote_average"]
+      voteAverage: json["vote_average"].toDouble(),
   );
 }
